#49: Added logic to SportCourtList Ui

+++ SportFinder/app/src/main/java/ru/moevm/sportfinder/screen/navigation/MainNavHost.kt
@@ -141,7 +141,7 @@ fun MainNavHost(
                 val state by viewModel.state.collectAsStateWithLifecycle()
                 updateBottomBarVisible(true)
                 updateTopBarType(false, null)
-                val navigateToSportCourtInfo = { courtId: Long ->
+                val navigateToSportCourtInfo = { courtId: Int ->
                     viewModel.onItemClicked(courtId)
                     navigationController.navigateToSportCourtInfo()
                 }
